gom-ai/requirements.txt
fastapi
uvicorn[standard]
Pillow
python-multipart
python-dotenv
openai
google-genai
httpx
tenacity
